action post dans like est comment validé

+++ backend/controlles/post.js
@@ -2,20 +2,22 @@ const models = require("../models");
 const fs = require("fs");
 
 exports.createPost = (req, res, next) => {
-  models.user.findOne({ where: { id: req.auth.userId } }).then((user) => {
-    models.post
-      .create({
-        userId: user.id,
-        content: req.body.content,
-        likes: 0,
-        image_url: req.file
-          ? `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
-          : null,
-      })
-      .then(() => res.status(201).json("Post created!!"));
-    //.catch((error) => res.status(404).json({ error }));
-  });
-  //.catch((error) => res.status(500).json({ error }));
+  models.user
+    .findOne({ where: { id: req.auth.userId } })
+    .then((user) => {
+      models.post
+        .create({
+          userId: user.id,
+          content: req.body.content,
+          likes: 0,
+          image_url: req.file
+            ? `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+            : null,
+        })
+        .then(() => res.status(201).json("Post created!!"))
+        .catch((error) => res.status(404).json({ error }));
+    })
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.updatePost = (req, res, next) => {
@@ -26,13 +28,15 @@ exports.updatePost = (req, res, next) => {
         return res.status(401).json({ error: "unauthorized request" });
       }
       if (req.file) {
-        const filename = post.image_url.split("/images")[1]; //le nom de l'image a supprimer
-        fs.unlink(`images/${filename}`, function (error) {
-          //supression de l'ancienne image
-          if (error) {
-            throw error;
-          }
-        });
+        if (post.image_url != null) {
+          const filename = post.image_url.split("/images")[1]; //le nom de l'image a supprimer
+          fs.unlink(`images/${filename}`, function (error) {
+            //supression de l'ancienne image
+            if (error) {
+              throw error;
+            }
+          });
+        }
       }
       const postObject = req.file
         ? {
@@ -53,16 +57,15 @@ exports.updatePost = (req, res, next) => {
 };
 
 exports.deletePost = (req, res, next) => {
-  models.post
-    .findOne({ where: { id: req.params.id } })
-    .then((post) => {
-      if (!post) {
-        return res.status(404).json({
-          error: new Error("post unexist"),
-        });
-      }
-      if (post.userId == req.auth.userId || req.auth.isAdmin == 1) {
-        //nom du fichier à supprimer
+  models.post.findOne({ where: { id: req.params.id } }).then((post) => {
+    if (!post) {
+      return res.status(404).json({
+        error: new Error("post unexist"),
+      });
+    }
+    if (post.userId == req.auth.userId || req.auth.isAdmin == 1) {
+      //nom du fichier à supprimer
+      if (post.image_url != null) {
         const filename = post.image_url.split("/images")[1];
         fs.unlink(`images/${filename}`, () => {
           post
@@ -75,19 +78,28 @@ exports.deletePost = (req, res, next) => {
             });
         });
       } else {
-        return res.status(401).json({ message: "Unauthorized request" });
+        post
+          .destroy({ where: { id: req.params.id } })
+          .then(() => {
+            res.status(201).json({ message: "Post delete !" });
+          })
+          .catch((error) => {
+            res.status(400).json({ error });
+          });
       }
-    })
-    .catch((error) => res.status(500).json({ error }));
+    } else {
+      return res.status(401).json({ message: "Unauthorized request" });
+    }
+  });
+  //.catch((error) => res.status(500).json({ error }));
 };
-
 exports.readAllPost = (req, res, next) => {
   models.post
     .findAll({
       include: [
         {
           model: models.user,
-          attributes: ["firstName", "lastName"],
+          attributes: ["firstName", "lastName", "image"],
         },
       ],
     })
